Added ability to expand one subject into multiple
* Added new value `expand_subjects` to `config.yml` into anketa.ais_user_source
which should specify subjects (based on their long code) that should be split
into multiple subjects and is defined in `config_local.yml`.
Each entry should contain new long code for the subject as well as new name of
the subject.

* In votr_runner.py we check if the subject is in list `expand_subjects`
and if so we multiply it based on this array defined in `config_local.yml`
If the subject was expanded we also add to subjects structure a value containing
its "old" name and code. This value is None if it is a normal subject.

* `AISRetriver` just obtains this new value passed from `AISUserSource` and
sends it into `runVotr()` function where this value is passed to `votr_runner.py`

* AISUserSource recieves this value from from `config.yml` passes it first to
`AISRetriver` and then if the results are concerning student it loads all
necessary subjects including the new ones added in `votr_runner.py`. These new
subjects are first added to all "Subject" tables and then Teacher-New_Subject
pairs are added to `TeachersSubjects` tables based on the "old" name and code
values send from `votr_runner` with expanded subjects.

Signed-off-by: Ondrej Jariabka <o.jariabka@gmail.com>